requirements.txt
configparser==3.5.0
cov-core==1.15.0
coverage==4.2
coveralls==1.1
nose-cov==1.6
nose2==0.6.5
python-daemon=2.1.1
slackclient==1.0.1
